% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{delete.intercept}
\alias{delete.intercept}
\title{Delete intercept from model matrix}
\usage{
delete.intercept(mm)
}
\arguments{
\item{mm}{Model matrix.}
}
\value{
A model matrix without intercept column.
}
\description{
A utility function to delete any intercept column from a model matrix, and
adjust the \code{"assign"} attribute correspondingly.  It is used by formula
handling functions like \code{mvr} and \code{model.matrix.mvr}.
}
\seealso{
\code{\link{mvr}}, \code{\link{model.matrix.mvr}}
}
\author{
Bjørn-Helge Mevik and Ron Wehrens
}
\keyword{internal}
