% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enter_confirm.R
\name{enter_confirm}
\alias{enter_confirm}
\title{Tools: "Are you sure?" Confirmation}
\usage{
enter_confirm()
}
\value{
A logical value, with \code{TRUE} as "Yes" and \code{FALSE} as "No"
}
\description{
Tools: "Are you sure?" Confirmation
}
\examples{
enter_confirm
}
\keyword{internal}
