% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enter_number.R
\name{enter_integer}
\alias{enter_integer}
\title{Tools: Enter integer}
\usage{
enter_integer(range = c(-Inf, Inf))
}
\arguments{
\item{range}{a vector with 2 numbers}
}
\value{
An integer
}
\description{
Tools: Enter integer
}
\examples{
enter_integer()
}
\keyword{internal}
