% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enter_interviewees.R
\name{enter_interviewee}
\alias{enter_interviewee}
\title{Tools: "How many interviewees?" Confirmation}
\usage{
enter_interviewee()
}
\value{
A list of numbers, from 1 to the designated amount
}
\description{
Tools: "How many interviewees?" Confirmation
}
\examples{
enter_interviewee
}
\keyword{internal}
