% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gasoline}
\alias{gasoline}
\title{Octane numbers and NIR spectra of gasoline}
\format{
A data frame with 60 observations on the following 2 variables.
\describe{ \item{octane}{a numeric vector.  The octane number.}
\item{NIR}{a matrix with 401 columns.  The NIR spectrum.} }
}
\source{
Kalivas, John H. (1997) Two Data Sets of Near Infrared Spectra
\emph{Chemometrics and Intelligent Laboratory Systems}, \bold{37}, 255--259.
}
\description{
A data set with NIR spectra and octane numbers of 60 gasoline samples.  The
NIR spectra were measured using diffuse reflectance as log(1/R) from 900 nm
to 1700 nm in 2 nm intervals, giving 401 wavelengths.  Many thanks to John
H. Kalivas.
}
\keyword{datasets}
