% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interview_1_1.R
\name{interview_1_1}
\alias{interview_1_1}
\title{Create an 1+1 interview list}
\usage{
interview_1_1(data_org)
}
\arguments{
\item{data_org}{A dataframe, containing column "category" and "word"}
}
\value{

}
\description{
A "1+1" interview refers to "1 chosen term, 1 non-chosen term". This is to ensure that interviewees are capable in handling non-major topics or seek help.
The dataframe inserted within this function must contain 2 columns, "category" and "word".
This function also returns a dataframe, that consist of all the terms choose by the interviewees.
}
\examples{
interview_1_1(df)
}
