% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interview_1n1.R
\name{interview_1n1}
\alias{interview_1n1}
\title{Create an 1+1 interview list}
\usage{
interview_1n1(data, i = NULL, seed = NULL)
}
\arguments{
\item{data}{Choices in dataframe, consisting of \code{category} and \code{word}}

\item{i}{List of Interviewee. By default (\code{NULL}), a pop-up will appear asking you how many interviewees there are, and will fill in accordingly as numbers.}

\item{seed}{Seed for the pseudo-random selection. By default (\code{NULL}), seed will be selected based on \code{Sys.time()}}
}
\value{
A dataframe, containing all terms choosen by the interviewees.
}
\description{
A "1+1" interview refers to "1 chosen term, 1 non-chosen term". This is to ensure that interviewees are capable of handling non-major topics or seeking help.
The dataframe inserted within this function must contain 2 columns, "category" and "word".
This function also returns a dataframe, that consists of all the terms chosen by the interviewees.
}
\examples{
interview_1n1(df)
}
