\name{pls-package}
\alias{pls-package}
\alias{pls}
\docType{package}
\title{
\packageTitle{pls}
}
\description{
\packageDescription{pls}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{pls}
\packageIndices{pls}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{pls}

Maintainer: \packageMaintainer{pls}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
