% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pls-package.R
\docType{package}
\name{pls}
\alias{pls}
\title{Partial Least Squares and Principal Component Regression}
\description{
Multivariate regression methods Partial Least Squares Regression (PLSR), Principal Component
Regression (PCR) and Canonical Powered Partial Least Squares (CPPLS).
}
