% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_category.R
\name{remove_category}
\alias{remove_category}
\title{Remove choices from list}
\usage{
remove_category(data, category)
}
\arguments{
\item{data}{Choices in dataframe, consisting of \code{category} and \code{word}}

\item{category}{Category to be removed, in character}
}
\value{
A dataframe, without the choices in \code{words}
}
\description{
Remove choices from list
}
\examples{
remove_category(data, words)
}
\keyword{internal}
