% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_choice.R
\name{remove_choice}
\alias{remove_choice}
\title{Remove choices from list}
\usage{
remove_choice(data, words)
}
\arguments{
\item{data}{Choices in dataframe, consisting of \code{category} and \code{word}}

\item{words}{Choices \emph{to be removed} in dataframe, consisting of \code{category} and \code{word}}
}
\value{
A dataframe, without the choices in \code{words}
}
\description{
Remove choices from list
}
\examples{
remove_choice(data, words)
}
\keyword{internal}
