% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_random.R
\name{sample_random}
\alias{sample_random}
\title{Sampling based on pseudo-random selection}
\usage{
sample_random(data, seed = NULL, force = FALSE)
}
\arguments{
\item{data}{Choices in dataframe, consisting of \code{category} and \code{word}}

\item{seed}{Seed for the pseudo-random selection. By default (\code{NULL}), seed will be selected based on \code{Sys.time()}}

\item{force}{Can the interview reject the choice? By default (\code{FALSE}), the interviewee \emph{can} reject the choice.}
}
\value{
A vector of 2 words, as follow
\itemize{
\item Category
\item Word
}
}
\description{
Sampling based on pseudo-random selection
}
\examples{
sample_random(data, seed = 1)
}
\keyword{internal}
