%% $Id$
\encoding{latin1}
\name{sensory}
\docType{data}
\alias{sensory}
\title{Multivariate Dataset of Sensory Data}
\description{
  Panel gives scores from a sensory panel on a set of 6
  attributes for 16 olive oil samples.  Quality gives the results of 5
  physico-chemical quality parameters.  The first five oils are Greek,
  the next five are Italian and the last six are Spanish.  This data set
  illustrates the qualitative capabilities of PLS.
}
\usage{data(sensory)}
\format{A data frame with matrices Quality (16 x 5) and Panel (16 x 6).}
\source{
  Massart, D. L., Vandeginste, B. G. M., Buydens, L. M. C., de Jong, S.,
  Lewi, P. J., Smeyers-Verbeke, J. (1998)  \emph{Handbook of
    Chemometrics and Qualimetrics: Part B}.  Elsevier. Tables 35.1 and 35.4.
}
\keyword{datasets}
