% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pls-package.R
\docType{package}
\name{pls}
\alias{pls-package}
\alias{pls}
\title{Partial Least Squares and Principal Component Regression}
\description{
Multivariate regression methods Partial Least Squares Regression (PLSR), Principal Component
Regression (PCR) and Canonical Powered Partial Least Squares (CPPLS).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/khliland/pls}
  \item Report bugs at \url{https://github.com/khliland/pls/issues}
}

}
\author{
\strong{Maintainer}: Kristian Hovde Liland \email{kristian.liland@nmbu.no}

Authors:
\itemize{
  \item Bjørn-Helge Mevik \email{b-h@mevik.net}
  \item Ron Wehrens
}

Other contributors:
\itemize{
  \item Paul Hiemstra [contributor]
}

}
\keyword{internal}
